Adding support for Validate-Job and Create-Job requests
This change also introduces the InterfaceManager class which is used to
manage a single ippusb interface. It keeps track of whether or not the
interface is currently receiving a chunked IPP message and holds a queue
which is used to hold respones to IPP requests.

BUG=chromium:885915
TEST=tested manually on device

Change-Id: I0f9e46fce22fcea443267ce610d05a36cc193ad3
Reviewed-on: https://chromium-review.googlesource.com/1450229
Commit-Ready: David Valleau <valleau@chromium.org>
Tested-by: David Valleau <valleau@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>