virtual-usb-printer: write document in IppManager
Previously, UsbPrinter assumed that if a chunked HTTP request completed,
the resulting message was a document being printed and should be written
to the output path.

Improve that logic a bit by passing the request body into IppManager. If
the request is a SendDocument request, IppManager will write the request
body to the document output path.

BUG=b:155432012
TEST=Unit tests

Change-Id: I85d28be8d3d53a79ced9922b9dfe1f4c9b3c82b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/virtual-usb-printer/+/2153769
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>