virtual-usb-printer: remove chunk-handling hack
Handling of chunked HTTP messages within UsbPrinter made the assumption
that the first chunk was always the IPP header and attributes. Now that
we are capable of directly removing the attributes, do so instead, and
remove this assumption.

We rework HandleChunkedIppUsbData to read the IppHeader only at the end
of chunk handling.

Because of this, we no longer need Get/SetDocument or
Get/SetChunkedIppHeader, so we can delete them.

BUG=chromium:1074523
BUG=b:155432012
TEST=tast run -checktestdeps=false -build=false ${DUT_IP} arc.Print
TEST=tast run -checktestdeps=false -build=false ${DUT_IP} \
     printer.PrintIPPUSB

Change-Id: Ifcba09fe4ddc798c6260dd66c4628d40ae14057c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/virtual-usb-printer/+/2166527
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Commit-Queue: Sean Kau <skau@chromium.org>