virtual-usb-printer: add HttpRequest class
Creates a new HttpRequest class in http_util.

The purpose of this class is to make parsing of incoming HTTP requests
more robust, and to allow us to read fields of requests.

For example, we need to know the target URI and request method to
properly route eSCL vs. IPP messages, and to report errors when we
receive a malformed message.

This patch only adds the new class and tests (as well as re-writing
RemoveHttpHeader to use the HttpRequest logic internally). Uses of the
old API functions will be converted in a later patch.

BUG=b:155432012
TEST=unit tests

Change-Id: I5111595e4402e99262a4a22b2fe4fd3af28cca59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/virtual-usb-printer/+/2153770
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Kalvin Lee <kdlee@chromium.org>