virtual-usb-printer: unbreak PrintUSB
A previous patch deleted the WriteDocument function in UsbPrinter,
replacing it with base::WriteFile. However, for larger print jobs
payloads for USB (not IPPUSB) printers, the payload may come in chunks,
and we need to write it progressively. WriteFile was completely
overwriting the file at each step, so it saved only the last chunk.

This was missed because only the PrintIPPUSB test was run with the new
changes.

Create a new static AppendToFile function in UsbPrinter which handles
the logic of creating and appending files, and use it in
UsbPrinter::HandleUsbData.

BUG=chromium:1077814
TEST=tast -verbose run -checktestdeps=false -build=false ${DUT} \
     printer.PrintUSB

Change-Id: Ib9cd637c76fcb290deb049b5c56324681c27c2e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/virtual-usb-printer/+/2180884
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Kalvin Lee <kdlee@chromium.org>
Reviewed-by: Sean Kau <skau@chromium.org>
Commit-Queue: Sean Kau <skau@chromium.org>