virtual-usb-printer: reduce logspam
When responding to a USB request with a queued message, log the size of
the sent message in bytes, and do not log an additional message
indicating that the leftover bytes have been re-queued.

This both adds more information, and allows for syslog to compress
consecutive messages, which greatly reduces logspam.

BUG=None
TEST=run scanner.ScanESCLIPP.network, look at /var/log/messages

Change-Id: Id5e05b2e392fb2f2ca5fd06af105de0bad8ace8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/virtual-usb-printer/+/2463888
Tested-by: Fletcher Woodruff <fletcherw@chromium.org>
Reviewed-by: Kalvin Lee <kdlee@chromium.org>
Commit-Queue: Fletcher Woodruff <fletcherw@chromium.org>